+++ ostd/specs/sync/examples/abstract_lock_tla.rs
