+++ ostd/specs/sync/examples/mod.rs
@@ -1,5 +1,5 @@
 #![allow(non_camel_case_types)]
 
-pub mod abstract_lock;
-pub mod mutex;
+pub mod abstract_lock_tla;
+pub mod mutex_tla;
 pub mod mutex_verussync;
